Rename from camel to Pascal case the left folders
