+++ src/App/index.js
@@ -1,16 +1,16 @@
-import * as database from "./database/api";
+import * as database from "../database/api";
 import localization, {
   addUpdatee as addLocalizationUpdatee,
-} from "./localization";
-import ConfirmDialog from "./ConfirmDialog/component";
-import Container from "./Container/component";
-import EmptyWorkspace from "./EmptyWorkspace/component";
-import Header from "./Header/component";
+} from "../localization";
+import ConfirmDialog from "../ConfirmDialog/component";
+import Container from "../Container/component";
+import EmptyWorkspace from "../EmptyWorkspace/component";
+import Header from "../Header/component";
 import React from "react";
-import Workspace from "./Workspace/component";
+import Workspace from "../Workspace/component";
 import atthemeEditorApi from "attheme-editor-api/browser";
-import parseTheme from "./parseTheme";
-import readFile from "./readFile";
+import parseTheme from "../parseTheme";
+import readFile from "../readFile";
 
 const HANDLE_SCROLL_INTERVAL = 200;
