+++ src/Dialog/component.js
