+++ src/Dialog/styles.scss
