+++ src/Dialog/variables.scss
