+++ src/Heading/component.js
