+++ src/Heading/styles.scss
