+++ src/Hint/component.js
