+++ src/Logo/component.js
