+++ src/Logo/styles.scss
