+++ src/Logo/variables.scss
