+++ src/Palettes/component.js
