+++ src/Palettes/styles.scss
