+++ src/Snackbar/component.js
