+++ src/Snackbar/styles.scss
