+++ src/Snackbar/variables.scss
