+++ src/Spinner/component.js
