+++ src/Spinner/styles.scss
