+++ src/Tabs/component.js
