+++ src/Tabs/styles.scss
