+++ src/Tabs/variables.scss
