+++ src/Variable/component.js
