+++ src/Variables/component.js
